add tests

+++ src/main/java/org/example/tasks/FirstTask.java
@@ -4,7 +4,7 @@ import static java.lang.Math.round;
 
 abstract class FirstTask {
 
-    static void getHello(final double number) {
+    static void getHello(final float number) {
         if (round(number) > 7) {
             System.out.println("Привет");
         }
