+++ src/main/java/org/example/tasks/ThirdTask.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 abstract class ThirdTask {
 
-    static void getNumbers(List<Double> numbers) {
+    static void getNumbers(List<Float> numbers) {
         for (double i : numbers) {
             if (i % 3 == 0) {
                 System.out.println("Число кратное трём: " + i);
